+++ laktory/models/resources/databricks/grant.py
@@ -8,28 +8,12 @@ from laktory.models.resources.pulumiresource import PulumiResource
 from laktory.models.resources.terraformresource import TerraformResource
 
 
-class GrantsGrant(BaseModel):
+class Grant(BaseModel, PulumiResource, TerraformResource):
     """
-    Grants grant
+    Databricks Grant
 
-    Attributes
-    ----------
-    principal:
-        User, group or service principal name
-    privileges:
-        List of allowed privileges
-    """
-
-    principal: str
-    privileges: list[str]
-
-
-class Grants(BaseModel, PulumiResource, TerraformResource):
-    """
-    Databricks Grants
-
-    Authoritative for all principals. Sets the grants of a securable and replaces any
-    existing grants defined inside or outside of Laktory.
+    Authoritative for a specific principal. Updates the grants of a securable to a
+    single principal. Other principals within the grants for the securables are preserved.
 
     Attributes
     ----------
@@ -37,12 +21,14 @@ class Grants(BaseModel, PulumiResource, TerraformResource):
         Name of the catalog to assign the grants to
     external_location:
         Name of the external location to assign the grants to
-    grants:
-        List of grant assigned to the selected object
     metastore:
         Name of the metastore to assign the grants to
     model
         Name of the user to assign the permission to.
+    principal:
+        User, group or service principal name
+    privileges:
+        List of allowed privileges
     schema:
         Name of the schema to assign the permission to.
     share:
@@ -59,18 +45,20 @@ class Grants(BaseModel, PulumiResource, TerraformResource):
     ```py
     from laktory import models
 
-    grants = models.resources.databricks.Grants(
+    grants = models.resources.databricks.Grant(
         catalog="dev",
-        grants=[{"principal": "metastore-admins", "privileges": ["CREATE_SCHEMA"]}],
+        principal="metastore-admins",
+        privileges=["CREATE_SCHEMA"],
     )
     ```
     """
 
     catalog: str = None
     external_location: str = None
-    grants: list[GrantsGrant]
     metastore: str = None
     model: str = None
+    principal: str
+    privileges: list[str]
     schema_: str = Field(
         None, validation_alias=AliasChoices("schema", "schema_")
     )  # required not to overwrite BaseModel attribute
@@ -89,7 +77,7 @@ class Grants(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def pulumi_resource_type(self) -> str:
-        return "databricks:Grants"
+        return "databricks:Grant"
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
@@ -101,7 +89,7 @@ class Grants(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def terraform_resource_type(self) -> str:
-        return "databricks_grants"
+        return "databricks_grant"
 
     @property
     def terraform_renames(self) -> dict[str, str]: