+++ laktory/models/resources/databricks/schema.py
@@ -4,7 +4,6 @@ from pydantic import model_validator
 
 from laktory.models.basemodel import BaseModel
 from laktory.models.grants.schemagrant import SchemaGrant
-from laktory.models.resources.databricks.grants import Grants
 from laktory.models.resources.databricks.table import Table
 from laktory.models.resources.databricks.volume import Volume
 from laktory.models.resources.pulumiresource import PulumiResource
@@ -24,8 +23,14 @@ class Schema(BaseModel, PulumiResource, TerraformResource):
         Text description of the catalog
     force_destroy:
         If `True` catalog can be deleted, even when not empty
+    grant:
+        Grant(s) operating on the Schema and authoritative for a specific principal.
+        Other principals within the grants are preserved. Mutually exclusive with
+        `grants`.
     grants:
-        List of grants operating on the schema
+        Grants operating on the Schema and authoritative for all principals.
+        Replaces any existing grants defined inside or outside of Laktory. Mutually
+        exclusive with `grant`.
     isolation_mode:
         Whether the catalog is accessible from all workspaces or a specific set
         of workspaces. Can be ISOLATED or OPEN. Setting the catalog to ISOLATED
@@ -63,6 +68,7 @@ class Schema(BaseModel, PulumiResource, TerraformResource):
     catalog_name: Union[str, None] = None
     comment: Union[str, None] = None
     force_destroy: bool = True
+    grant: Union[SchemaGrant, list[SchemaGrant]] = None
     grants: list[SchemaGrant] = None
     name: str
     storage_root: str = None
@@ -117,17 +123,7 @@ class Schema(BaseModel, PulumiResource, TerraformResource):
         resources = []
 
         # Schema grants
-        if self.grants:
-            resources += [
-                Grants(
-                    resource_name=f"grants-{self.resource_name}",
-                    schema=f"${{resources.{self.resource_name}.id}}",
-                    grants=[
-                        {"principal": g.principal, "privileges": g.privileges}
-                        for g in self.grants
-                    ],
-                )
-            ]
+        resources += self.get_grants_additional_resources()
 
         if self.volumes:
             for v in self.volumes:
@@ -149,7 +145,7 @@ class Schema(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["tables", "volumes", "grants"]
+        return ["tables", "volumes", "grant", "grants"]
 
     # ----------------------------------------------------------------------- #
     # Terraform Properties                                                    #