+++ laktory/models/resources/databricks/table.py
@@ -7,7 +7,6 @@ from laktory._logger import get_logger
 from laktory.models.basemodel import BaseModel
 from laktory.models.grants.tablegrant import TableGrant
 from laktory.models.resources.baseresource import ResourceLookup
-from laktory.models.resources.databricks.grants import Grants
 from laktory.models.resources.pulumiresource import PulumiResource
 from laktory.models.resources.terraformresource import TerraformResource
 
@@ -72,8 +71,14 @@ class Table(BaseModel, PulumiResource, TerraformResource):
         External tables are supported in multiple data source formats. The string constants identifying these formats
         are DELTA, CSV, JSON, AVRO, PARQUET, ORC, TEXT. Change forces creation of a new resource. Not supported for
         MANAGED tables or VIEW.
+    grant:
+        Grant(s) operating on the Table and authoritative for a specific principal.
+        Other principals within the grants are preserved. Mutually exclusive with
+        `grants`.
     grants:
-        List of grants operating on the schema
+        Grants operating on the Table and authoritative for all principals.
+        Replaces any existing grants defined inside or outside of Laktory. Mutually
+        exclusive with `grant`.
     lookup_existing:
         Specifications for looking up existing resource. Other attributes will
         be ignored.
@@ -118,6 +123,7 @@ class Table(BaseModel, PulumiResource, TerraformResource):
     columns: Union[list[TableColumn], None] = None
     comment: Union[str, None] = None
     data_source_format: str = "DELTA"
+    grant: Union[TableGrant, list[TableGrant]] = None
     grants: list[TableGrant] = None
     lookup_existing: TableLookup = Field(None, exclude=True)
     name: str
@@ -194,18 +200,9 @@ class Table(BaseModel, PulumiResource, TerraformResource):
         """
         resources = []
 
-        # Schema grants
-        if self.grants:
-            resources += [
-                Grants(
-                    resource_name=f"grants-{self.resource_name}",
-                    table=f"${{resources.{self.resource_name}.id}}",
-                    grants=[
-                        {"principal": g.principal, "privileges": g.privileges}
-                        for g in self.grants
-                    ],
-                )
-            ]
+        # Table grants
+        resources += self.get_grants_additional_resources()
+
         return resources
 
     # ----------------------------------------------------------------------- #
@@ -219,6 +216,7 @@ class Table(BaseModel, PulumiResource, TerraformResource):
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
         return [
+            "grant",
             "grants",
         ]