+++ tests/test_table.py
@@ -66,6 +66,7 @@ def test_model():
         ],
         "comment": None,
         "data_source_format": "DELTA",
+        "grant": None,
         "grants": None,
         "name": "slv_stock_prices",
         "properties": None,